[PIO] GZip and copy files to build_output directory

+++ tools/pio/gzip-firmware.py
@@ -1,4 +1,4 @@
-# Original: https://github.com/arendst/Tasmota/blob/development/pio/gzip-firmware.py
+# Inspired by: https://github.com/arendst/Tasmota/blob/development/pio/gzip-firmware.py
 # Thanks Theo & Jason2866 :)
 
 Import('env')
@@ -9,18 +9,19 @@ import gzip
 OUTPUT_DIR = "build_output{}".format(os.path.sep)
 
 def bin_gzip(source, target, env):
-    variant = str(target[0]).split(os.path.sep)[2]
-    
-    # create string with location and file names based on variant
-    bin_file = "{}firmware{}{}.bin".format(OUTPUT_DIR, os.path.sep, variant)
-    gzip_file = "{}firmware{}{}.bin.gz".format(OUTPUT_DIR, os.path.sep, variant)
+    bin_file = str(target[0])
+    #print("\u001b[33m bin_file: \u001b[0m  {}".format(bin_file))
+    if ".bin" in bin_file:
+        # create string with location and file names based on variant
+        gzip_file = "{}.gz".format(bin_file)
 
-    # check if new target files exist and remove if necessary
-    if os.path.isfile(gzip_file): os.remove(gzip_file)
+        # check if new target files exist and remove if necessary
+        if os.path.isfile(gzip_file): os.remove(gzip_file)
 
-    # write gzip firmware file
-    with open(bin_file,"rb") as fp:
-        with gzip.open(gzip_file, "wb") as f:
-            shutil.copyfileobj(fp, f)
+        # write gzip firmware file
+        with open(bin_file,"rb") as fp:
+            with gzip.open(gzip_file, "wb") as f:
+                shutil.copyfileobj(fp, f)
+
 
 env.AddPostAction("$BUILD_DIR/${PROGNAME}.bin", [bin_gzip])
